Moved BootstrapButton to separate file; added props validation
